% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.partial.pairwise.R
\name{cor.partial.pairwise}
\alias{cor.partial.pairwise}
\title{Calculate pairwise partial correlations}
\usage{
\code{
cor.partial.pairwise(x, method="pearson", ...)}
}
\arguments{
\item{x}{a numeric matrix. Each row should correspond to a sample, with variables in columns.}

\item{...}{additional arguments passed to \code{cor}.}

\item{verbose}{logical, whether to output status of calculations.}

\item{method}{a string, the correlation coefficient to be computed; passed to \code{cor}.}
}
\value{
a numeric matrix containing the pairwise partial correlations of variables in \code{x}. Elements above and below the diagonal are symmetric.
}
\description{
This function calculates partial correlations for all pairwise combinations of variables in a matrix,
while controlling for variation explained by all other elements in the matrix
}
