% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.partial.R
\name{cor.partial}
\alias{cor.partial}
\title{Calculate partial correlations}
\usage{
\code{
cor.partial(x, y)}
}
\arguments{
\item{x}{a numeric vector or matrix. If a vector, each element should correspond to a sample. If a matrix, each row should correspond to a sample, with variables in columns.}

\item{y}{a numeric matrix. Each row should correspond to a sample; each column should correspond to a variable.}

\item{...}{additional arguments passed to \code{cor}.}

\item{verbose}{logical, whether to output status of calculations.}
}
\value{
a numeric matrix containing the partial correlations of variables in \code{x} and \code{y}. Rows in the result correspond to variables in \code{x}; columns in the result correspond to variables in \code{y}.
}
\description{
This function calculates partial correlations between a vector or matrix and a second matrix,
while controlling for variation explained by other elements in the second matrix.
}
\details{
If only \code{x} is provided, the pairwise partial correlations among all columns are computed. If \code{y} is provided, the pairwise partial correlations between each column of \code{x} and each column of {y} are computed.
}
