% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statCors.R
\name{statCors}
\alias{statCors}
\title{Calculate statistics on a correlation matrix}
\usage{
\code{
statCors(cors, margin=1,
         var_threshold=0, sum_abs_threshold=0,
         min_max_threshold=0,
         all_thresholds=FALSE,
         return_order="var",
         )}
}
\arguments{
\item{cors}{a numeric matrix of correlations.}

\item{margin}{the dimension to apply statistics over; passed to \code{apply}. \code{1} indicates rows, \code{2} indicates columns.}

\item{var_threshold}{numeric, the threshold for the variance across variables. High values filter to include only variables that have widely-distributed correlations with other variables. Defaults to 0, which}

\item{sum_abs_threshold}{numeric, the threshold for the minimum sum of the absolute values of correlations across all other variables. High values filter to include only variables that have large magnitude correlations (positive or negative) with many other variables.}

\item{min_max_threshold}{numeric, the absolute value of the minimum/maximum correlations for inclusion of variable. This threshold is met if any correlation is \code{<= -min_max_threshold} or \code{>= min_max_threshold}}

\item{all_thresholds}{boolean, whether all thresholds must be met for variables to be included in the result. By default, variables that meet any of the thresholds are included.}

\item{return_order}{string, the variable to sort results by. Can be any of "var", "sum_abs", "min", "max", or unique partial matches. By default sort is decreasing, except for "min", where it is increasing. Defaults to "var".}
}
\value{
a data frame containing the summary statistics of correlations for each variable.
}
\description{
This function calculates some summary statistics on a correlation matrix. It was designed to be
used with a matrix of correlations of genes x some other set of variables (e.g. cytokine assay
values).
}
